Improve help messaging
Running with no argument (`plz`) now prints a more simplistic usage
explanation (not generated from ArgParse, because that automatically
generated help/usage text does not match well with the subcommand format
in use here). It also prints a list of available commands read from the
config file, instead of simply referring the user to the config file.

Overrides `plz help` to do the same as `plz`, as above.

Finally, it prints the list of available commands (but no usage help)
when an invalid command is issued.

Resolves: #6